# Greedy object hash