@corka149 2025-07-20
Sounds very interesting but I am far away from a computer for two weeks. 😬 So, when feedback takes some time.