@Valian 2025-07-28
@corka149 successfully rebased and added additional tests for ancestor path and object_hash. Everything seems fine, just credo checks are failing - for some reason I'm unable to run them locally, that's why I'll need a few iterations to have it sorted out. 

It would be great if you could check it - implementation should be quite straightforward, even if looking a bit intimidating ;)